requirements.txt
streamlit
langchain
langchain_groq
langchain_core
langchain_community
langchain_experimental
python_dotenv
pandas
numpy
scikit-learn